Rename JS file to relate more to contents.
